refactor(transform): split unroll.rs to satisfy the 2000-line file gate
The #2308 fix pushed crates/perry-transform/src/unroll.rs to 2032 lines,
tripping the lint job's file-size gate (./scripts/check_file_size.sh, cap
2000). Move the file to unroll/mod.rs and extract the #2308 escaping-id
analysis (compute_loop_escaping_ids + its count/declared-id helpers) into a
sibling unroll/escape_analysis.rs, re-exported via an explicit named use.
No behavior change — mod.rs is now 1801 lines, escape_analysis.rs 245.
